Apartment_research
Research of ads for the sale of apartments
We have the data of the Yandex.Realty service - an archive of ads for the sale of apartments in St. Petersburg and neighboring settlements for several years. My job is to set the parameters. This will allow to build an automated system: it will track anomalies and fraudulent activity.

Two types of data are available for each apartment for sale. The first one is entered by the user, the second one is obtained automatically on the basis of cartographic data. For example, the distance to the center, airports, nearest parks.

Objectives of the study: to determine which parameters affect the pricing of an apartment. Determine what types of dependencies exist, which ones are strong and which are weak.

Research project steps:
1. pre-process the data, identify anomalous values ​​and remove them. Clear data from obvious duplicates, check the data type and change it if necessary.
2. The next step is to define new variables: cost per square meter, month, year and day of the week.
3. Then, conduct a self-study and determine what parameters affect the cost of the apartment. Compare the dependence among themselves, which of the parameters of the apartment is decisive in pricing.